build go, and then build minikube
Also, apply a patch to prevent MakeDiskImage from chowing, which
fails under confinement.

Note that this still will not work because kvm2 tries to figure
out the IP address by reading the DHCP leases file, which is
not accessible while under confinement.

Also, remove the home plug, since it is unneeded, and update
to current snapcraft.yaml standards.